requirements.txt
asgiref==3.8.1
Django==5.2
pillow==11.2.1
sqlparse==0.5.3
tzdata==2025.2
pytest==8.3.3
pytest-django==4.9.0
factory-boy==3.3.1
faker==26.0.0
coverage==7.6.1
